Initial commit
This is an initial example which we can build on to add additional
host and embedded languages. For now there is just a single host
language (C++) and a single embedded one (json).
The extension's structure mostly comes from the Yeoman VSCode
extension generator (`yo code`), and the contents are based on
sample code from Microsoft
[https://code.visualstudio.com/api/language-extensions/syntax-highlight-guide]
and additional examples such as
[https://github.com/mjbvz/vscode-fenced-code-block-grammar-injection-example]